add doashboard redux slice and dummy course service
