+++ course-registration-assistant-app/src/app/store.js
@@ -2,10 +2,11 @@ import { configureStore } from '@reduxjs/toolkit';
 
 // Reducers
 import loginReducer from '../containers/Login/loginSlice';
-
+import dashboardReducer from '../containers/Dashboard/DashboardSlice';
 
 export default configureStore({
   reducer: {
-    login: loginReducer
+    dashboard: dashboardReducer,
+    login: loginReducer,
   }
 });
