+++ course-registration-assistant-app/src/containers/Dashboard/Dashboard.js
@@ -1,10 +1,11 @@
-import React from 'react';
-import { useDispatch } from 'react-redux';
+import React, { useEffect } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
 import { useHistory } from 'react-router';
 import HeaderNla25 from '../../components/HeaderNla25/HeaderNla25';
 import Week from '../../components/Week';
 import { userLogout } from '../Login/loginSlice';
 import style from './Dashboard.module.scss';
+import { getAllCourses } from './DashboardSlice';
 
 const DAY_MAX_WORKING_TIME_BLOCK_BK = 24;
 const DAY_MAX_WORKING_TIME_BLOCK = 13;
@@ -15,145 +16,19 @@ const timeTableLabel = ["7:00", "8:00", "9:00", "10:00","11:00", "12:00", "13:00
 const weekDays = ["SUN", "MON", "TUE", "WED", "THI", "FRI", "SAT"];
 
 
-const initData = [
-  {
-      "name": "Monday",
-      "courses": []
-  },
-  {
-      "name": "Tuesday",
-      "courses": [
-          // [
-          //     {
-          //         id: "1",
-          //         title: "2200",
-          //         details: { instructor: "Rahim", credit: 3 },
-          //         offset: 3,
-          //         duration: 4,
-          //         preCoursePos: 0,
-          //     },
-          //     {
-          //         id: "2",
-          //         title: "2700",
-          //         details: { instructor: "Michael", credit: 3 },
-          //         offset: 7,
-          //         duration: 4,
-          //         preCoursePos: 5,
-          //     },
-          // ],
-          // [
-          //     {
-          //         id: "3",
-          //         title: "2300",
-          //         details: { instructor: "Brian", credit: 3 },
-          //         offset: 3,
-          //         duration: 4,
-          //         preCoursePos: 0,
-          //     },
-          // ],
-          // [
-          //     {
-          //         id: "4",
-          //         title: "3380",
-          //         details: { instructor: "Gabriel", credit: 3 },
-          //         offset: 3,
-          //         duration: 4,
-          //         preCoursePos: 0,
-          //     },
-          // ]]
-          [
-              {
-                  id: "1",
-                  title: "2200",
-                  details: { instructor: "Rahim Rahim", credit: 3 },
-                  offset: 3,
-                  duration: 4,
-                  preCoursePos: 0,
-              },
-              
-              {
-                  id: "3",
-                  title: "2300",
-                  details: { instructor: "Brian Vo", credit: 3 },
-                  offset: 3,
-                  duration: 4,
-                  preCoursePos: 0,
-              },
-              {
-                  id: "4",
-                  title: "3380",
-                  details: { instructor: "Gabriel", credit: 3 },
-                  offset: 3,
-                  duration: 4,
-                  preCoursePos: 0,
-              },
-          ],
-          [
-              {
-                  id: "2",
-                  title: "2700",
-                  details: { instructor: "Michael", credit: 3 },
-                  offset: 7,
-                  duration: 4,
-                  preCoursePos: 5,
-              },
-          ],
-          [
-              
-          ]]
-  },
-  {
-      "name": "Wednesday",
-      "courses": [
-          [
-              {
-                  id: "5",
-                  title: "2200",
-                  details: { instructor: "Rahim", credit: 3 },
-                  offset: 3,
-                  duration: 4,
-                  preCoursePos: 0,
-              },
-              {
-                  id: "6",
-                  title: "2700",
-                  details: { instructor: "Michael", credit: 3 },
-                  offset: 5,
-                  duration: 4,
-                  preCoursePos: 5,
-              },
-          ],
-          [
-              {
-                  id: "8",
-                  title: "3380",
-                  details: { instructor: "Gabriel", credit: 3 },
-                  offset: 10,
-                  duration: 6,
-                  preCoursePos: 0,
-              },
-          ]]
-  },
-  {
-      "name": "Thirsday",
-      "courses": []
-  },
-  {
-      "name": "Friday",
-      "courses": []
-  },
-  {
-      "name": "Saturday",
-      "courses": []
-  },
-];
+
 
 const Dashboard = () => {
 
+  const dispatch = useDispatch();
+  const courses = useSelector(state => state.dashboard.courses);
 
+  useEffect(() => {
+    dispatch(getAllCourses());
+  }, []);
 
   const history = useHistory();
-  const dispatch = useDispatch();
+
   const logout = () => {
     dispatch(userLogout());
     history.replace("/login");
@@ -172,7 +47,7 @@ const Dashboard = () => {
             This is sidebar
             </div>
           <div className="col-9">
-            <Week timeFrames={timeTableLabel} days={initData}/>
+            <Week timeFrames={timeTableLabel} days={courses}/>
           </div>
         </div>
         <div className="row">
@@ -182,9 +57,6 @@ const Dashboard = () => {
           </div>
         </div>
       </div>
-
-
-
     </div>
   )
 }
