+++ course-registration-assistant-app/src/services/CourseService/index.js
@@ -1,0 +1,1 @@
+export * as default from './CourseService';
